requirements.txt
transnetv2-pytorch
opencv-python
numpy
qdrant-client
yt-dlp
sqlalchemy
psycopg2-binary
alembic
celery
redis
transformers
Pillow
grpcio
grpcio-tools
sentencepiece
